requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
langchain-anthropic>=0.1.1
chromadb>=0.4.22
boto3>=1.34.0
pandas>=2.1.4
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.66.1
rich>=13.7.0
pypdf>=4.0.0
pydantic>=2.5.0
aiohttp>=3.9.1
tenacity>=8.2.3
typing-extensions>=4.9.0
colorama>=0.4.6  # For Windows color support
watchdog>=3.0.0  # For file monitoring